Get memes and append to DOM
Created new memes folder in assets/scripts with a file for events.js,
api.js, and ui.js. Created a 'show me the memes' button that calls
on the getMyMemes function in api.js. API makes a GET request for all
memes from the user.

Created handlebars file assets/scripts/templates/meme-stream.handlebars
that takes each meme from the meme array and appends it to the DOM with
the image, text_1 and text_2.

Added CSS styling in assets/styles/index.scss to format the memes with
the top and bottom text, with images scaling based on the width of the
viewport.