Updated handlebars template
Meme-stream handlebars template updated to create html to input data
to update each meme. Data-id is assigned equal to the meme-id. Need
to update the updatePreview function to correspond the id of the image
with the id of the meme so each meme can be updated and previewed in
real time.