Created comments section for each meme
Added section in meme-stream.handlebars file to display all comments
for each meme. Including a form to input a new comment with a
 'create comment' button to submit the form.

Created function to get form data and submit a POST request to the
api to create a new comment, then update the meme-stream with the
updated data.